+++ inc/carousel-slider/class-shapla-carousel-slider.php
@@ -48,23 +48,23 @@ if ( ! class_exists( 'Shapla_Carousel_Slider' ) ) {
 		public function carousel_slider_post() {
 			$blog = new Shapla_Blog();
 			?>
-			<div class="blog-grid-inside">
+            <div class="blog-grid-inside">
 				<?php $blog->post_thumbnail(); ?>
-				<header class="entry-header">
+                <header class="entry-header">
 					<?php
 					$blog->post_category();
 					$blog->post_title();
 					?>
-				</header>
-				<div class="entry-summary"><?php echo get_the_excerpt(); ?></div>
+                </header>
+                <div class="entry-summary"><?php echo get_the_excerpt(); ?></div>
 				<?php $blog->post_tag(); ?>
-				<footer class="entry-footer">
+                <footer class="entry-footer">
 					<?php
 					$blog->post_author();
 					$blog->post_date();
 					?>
-				</footer>
-			</div>
+                </footer>
+            </div>
 			<?php
 		}
 
@@ -76,14 +76,17 @@ if ( ! class_exists( 'Shapla_Carousel_Slider' ) ) {
 		 * @return mixed
 		 */
 		public function default_settings( $options ) {
-			$options['product_title_color']       = shapla_default_options()->heading_color;
-			$options['product_button_bg_color']   = shapla_default_options()->primary_button->background;
-			$options['product_button_text_color'] = shapla_default_options()->primary_button->text;
-			$options['nav_color']                 = shapla_adjust_color_brightness(
-				shapla_default_options()->primary_button->text,
-				- 14
-			);
-			$options['nav_active_color']          = shapla_default_options()->primary_button->background;
+			$btn_bg_default   = shapla_default_options()->primary_button->background;
+			$btn_text_default = shapla_default_options()->primary_button->text;
+			$heading_color    = get_theme_mod( 'typography_heading_color', shapla_default_options()->heading_color );
+			$btn_bg           = get_theme_mod( 'button_primary_background_color', $btn_bg_default );
+			$btn_text         = get_theme_mod( 'button_primary_text_color', $btn_text_default );
+
+			$options['product_title_color']       = $heading_color;
+			$options['product_button_bg_color']   = $btn_bg;
+			$options['product_button_text_color'] = $btn_text;
+			$options['nav_color']                 = shapla_adjust_color_brightness( $btn_text, - 14 );
+			$options['nav_active_color']          = $btn_bg;
 			$options['margin_right']              = 30;
 			$options['lazy_load_image']           = 'on';